separate input_selection from fee
